Rename is dumpnodes
